+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added option to hide the Skyblock Level from the chat messages,
 - Added option to change the way the Skyblock Level gets displayed in the chat (only working when skyblock level and fancy player message format are enabled)
 - Added highlight the voidling extremist in pink color
+- Added highlight corrupted mobs in purple color
 
 ## Version 0.6 - Ashfang and Summoning Mobs